+++ yash-builtin/src/jobs.rs
@@ -89,14 +89,12 @@ use crate::common::report_failure;
 use crate::common::syntax::parse_arguments;
 use crate::common::syntax::Mode;
 use crate::common::syntax::OptionSpec;
-use std::fmt::Write;
 use yash_env::builtin::Result;
+use yash_env::job::fmt::Accumulator;
 use yash_env::job::id::parse;
 use yash_env::job::id::parse_tail;
 use yash_env::job::id::FindError;
-use yash_env::job::Job;
 use yash_env::semantics::Field;
-use yash_env::system::SharedSystem;
 use yash_env::Env;
 use yash_syntax::source::pretty::Annotation;
 use yash_syntax::source::pretty::AnnotationType;
@@ -108,48 +106,6 @@ const OPTIONS: &[OptionSpec] = &[
     OptionSpec::new().short('l').long("verbose"),
     OptionSpec::new().short('p').long("pgid-only"),
 ];
-
-struct Accumulator {
-    current_job_index: Option<usize>,
-    previous_job_index: Option<usize>,
-    show_pid: bool,
-    pgid_only: bool,
-    print: String,
-    indices_reported: Vec<usize>,
-}
-
-impl Accumulator {
-    /// Processes one job.
-    ///
-    /// 1. Formats a job report in `self.print` so it can be printed later.
-    /// 1. Remembers the job index in `self.indices_reported` so the job can be
-    ///    removed later.
-    fn report(&mut self, index: usize, job: &Job, system: &SharedSystem) {
-        use yash_env::job::fmt::{Marker, Report, State};
-        let report = Report {
-            number: index + 1,
-            marker: if self.current_job_index == Some(index) {
-                Marker::CurrentJob
-            } else if self.previous_job_index == Some(index) {
-                Marker::PreviousJob
-            } else {
-                Marker::None
-            },
-            pid: self.show_pid.then_some(job.pid),
-            state: State::from_process_state(job.state, system),
-            name: &job.name,
-        };
-
-        if self.pgid_only {
-            writeln!(self.print, "{}", job.pid)
-        } else {
-            writeln!(self.print, "{report}")
-        }
-        .unwrap();
-
-        self.indices_reported.push(index);
-    }
-}
 
 fn find_error_message(error: FindError, operand: &Field) -> Message {
     Message {
@@ -192,14 +148,14 @@ pub async fn main(env: &mut Env, args: Vec<Field>) -> Result {
     if operands.is_empty() {
         // Report all jobs
         for (index, job) in &env.jobs {
-            accumulator.report(index, job, &env.system)
+            accumulator.add(index, job, &env.system)
         }
     } else {
         // Report jobs specified by the operands
         for operand in operands {
             let job_id = parse(&operand.value).unwrap_or_else(|_| parse_tail(&operand.value));
             match job_id.find(&env.jobs) {
-                Ok(index) => accumulator.report(index, &env.jobs[index], &env.system),
+                Ok(index) => accumulator.add(index, &env.jobs[index], &env.system),
                 Err(error) => {
                     return report_failure(env, find_error_message(error, &operand)).await
                 }